--- lib/presentation/pages/main/community/community_body.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CommunityBody extends StatelessWidget {
-  CommunityBody({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Text('communityScreen'),
-    );
-  }
-}
